Seed functions updated
detectPlanet(radius) is a function that calculates the x and y of the
point that the seed will hit the edge of the planet.

stopMovement() utilizes detectPlanet(r) to stop the movement of the seed
when it gets within the range of the radius. Currently does not work

sketch.js updated for testing purposes